BikeCsvImporter: refactor by splitting into parts
